Fix: Rename components "Try" for "TryBtn"
